Refactor code to split App component
